Use a custom menu item for application items
This introduces ImAppMenuItem: a menu item which shows a small triangle next to
an application's name if the associated app is running.

The running state is communicated to the menu by giving the "launch" action a
boolean state.

This depends on a patch to gtk which creates custom menu items from
gtk_menu_new_from_model when the x-canonical-type attribute is set on a menu
item in the model.