Use indicator-message icons with chat status
Changes the state of the "messages" action from a boolean (draws-attention) to
a string (icon-name).  This has the added benefit that more logic is moved from
the plugin into the service.

It also fixes an edge case: the messaging menu didn't have the blue icon after
the service restarted (if anything was drawing attention).