Move service .xml into a common directory
Both the service (in src/) and the client library (in libmessaging-menu/) need
access to the dbus interface description file.  Until now, it resided in src,
with both Makefiles calling gdbus-codegen on it.

This patch moves the file to common/ and builds a convenience library that
contains only the generated code.